conversion/requirements.txt
pybboxes
Pillow